hwmon: (corsair_cpro) Use a separate buffer for sending commands
Introduce cmd_buffer, a separate buffer for storing only
the command that is sent to the device. Before this separation,
the existing buffer was shared for both the command and the
report received in ccp_raw_event(), which was copied into it.

However, because of hidraw, the raw event parsing may be trigerred
in the middle of sending a command, resulting in outputting gibberish
to the device.

Signed-off-by: Aleksa Savic <savicaleksa83@gmail.com>